Tab or snippets?
We often use tab to select a completion item, cycle between the
candidates. With respect to completion plugin alone, it's simple, when
pupup menu is visible, replace <tab> with <c-n>, <s-tab> with <c-p>.
However, when it comes to the snippets plugin, things become
a little bit complicated. Only auto completion plugins are considered for
now.